Updated the .gitignore file to include JSON files for project debugging on development environment. The initial files for debugging are supplied with their default values (can be changed). Moved the TCPServerSync class in a separate namespace which is now being used by the program. Upgraded the application to use dotnet core v2.2.
