+++ .gitignore
@@ -30,7 +30,9 @@ bld/
 # Visual Studio 2015/2017 cache/options directory
 .vs/
 # Visual Studio Code directory
-.vscode/
+.vscode/*
+!.vscode/tasks.json
+!.vscode/launch.json
 # Uncomment if you have tasks that create the project's static files in wwwroot
 #wwwroot/
