+++ vlab-docker-dotnet-core-stream-app/vlab-docker-dotnet-core-stream-app.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.1</TargetFramework>
+    <TargetFramework>netcoreapp2.2</TargetFramework>
     <RootNamespace>vlab_docker_dotnet_core_stream_app</RootNamespace>
   </PropertyGroup>
